moved event to beat
